requirements.txt
asgiref==3.3.1
Django==3.1.7
django-choices==1.7.1
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-font-icons==1.0.1
django-js-asset==1.2.2
Pillow==8.1.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
